Validate module support
