+++ javaparser-testing/src/test/java/com/github/javaparser/ast/validator/Java1_1ValidatorTest.java
@@ -111,26 +111,6 @@ public class Java1_1ValidatorTest {
     }
 
     @Test
-    public void moduleRequires() {
-        ParseResult<CompilationUnit> result = javaParser.parse(COMPILATION_UNIT, provider("module x{requires " + allModifiers + " a;}"));
-        assertProblems(result,
-                "(line 1,col 10) Can have only one of 'public', 'protected', 'private'.",
-                "(line 1,col 10) Can have only one of 'final', 'abstract'.",
-                "(line 1,col 10) 'transient' is not allowed here.",
-                "(line 1,col 10) 'volatile' is not allowed here.",
-                "(line 1,col 10) 'final' is not allowed here.",
-                "(line 1,col 10) 'synchronized' is not allowed here.",
-                "(line 1,col 10) 'default' is not allowed here.",
-                "(line 1,col 10) 'native' is not allowed here.",
-                "(line 1,col 10) 'private' is not allowed here.",
-                "(line 1,col 10) 'protected' is not allowed here.",
-                "(line 1,col 10) 'strictfp' is not allowed here.",
-                "(line 1,col 10) 'abstract' is not allowed here.",
-                "(line 1,col 10) 'public' is not allowed here."
-        );
-    }
-
-    @Test
     public void constructor() {
         ParseResult<CompilationUnit> result = javaParser.parse(COMPILATION_UNIT, provider("class X{" + allModifiers + "X(){};}"));
         assertProblems(result,
